Refactor: V-1.6.1 Move old presentation layer to `presentation.old`
This commit moves the existing Fragment-based presentation layer components (`PlayerInfoBottomSheet.kt`, `RankingFragment.kt`) into a new package `ir.miare.androidcodechallenge.presentation.old`.

This is likely in preparation for a new presentation layer implementation.

Key changes:
- Renamed package for `PlayerInfoBottomSheet.kt` and `RankingFragment.kt`.
- Updated `tools:context` in `fragment_ranking.xml` to reflect the new package name.
- Updated app version to 1.6.1 (version code 13).
