+++ app/src/main/java/ir/miare/androidcodechallenge/presentation/old/PlayerInfoBottomSheet.kt
@@ -1,4 +1,4 @@
-package ir.miare.androidcodechallenge.presentation
+package ir.miare.androidcodechallenge.presentation.old
 
 import android.os.Bundle
 import android.view.LayoutInflater
