+++ app/src/main/java/ir/miare/androidcodechallenge/presentation/old/RankingFragment.kt
@@ -1,4 +1,4 @@
-package ir.miare.androidcodechallenge.presentation
+package ir.miare.androidcodechallenge.presentation.old
 
 import android.os.Bundle
 import android.os.Handler
