+++ app/src/main/res/layout/fragment_ranking.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.RankingFragment">
+    tools:context=".presentation.old.RankingFragment">
 
     <ScrollView
         android:layout_width="0dp"
